WIP in translate command.
Actually we insert the translation, choose de language and after update the translation with the language chosen.
Then show a list of inline buttons representing all the words in the text to translate.

We defined and created on config.py translatios_details table where all the query to de API of all the words of the text will be saved.